restructure styles directories
