+++ craco.config.js
@@ -2,11 +2,12 @@ const path = require('path');
 module.exports = {
   webpack: {
     alias: {
-      '@': path.resolve(__dirname, 'src'),
       '@components': path.resolve(__dirname, 'src/components'),
       '@providers': path.resolve(__dirname, 'src/providers'),
       '@features': path.resolve(__dirname, 'src/features'),
       '@assets': path.resolve(__dirname, 'src/assets'),
+      '@styles': path.resolve(__dirname, 'src/styles'),
+      '@': path.resolve(__dirname, 'src'),
     },
   }
 };
